+++ pyunity/values/__init__.py
@@ -4,14 +4,11 @@
 
 __all__ = ["Clock", "ImmutableStruct", "Quaternion"]
 from .vector import __all__ as _vector_all
-from .texture import __all__ as _texture_all
 from .abc import __all__ as _abc_all
 __all__.extend(_vector_all)
-__all__.extend(_texture_all)
 __all__.extend(_abc_all)
 
 from .vector import *
-from .texture import *
 from .quaternion import Quaternion
 from .other import Clock, ImmutableStruct
 from .abc import *
