+++ tests/pyunity/testMeshes/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (c) 2020-2022 The PyUnity Team
+# This file is licensed under the MIT License.
+# See https://docs.pyunity.x10.bz/en/latest/license.html
+
+__all__ = ["TestCase", "almostEqual"]
+from .. import TestCase, almostEqual
